flaskr/requirements.txt
attrs==22.2.0
cachelib==0.10.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
coverage==7.0.5
Flask==2.1.0
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-Sessions==0.1.5
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
pluggy==0.13.1
protobuf==4.22.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytest==6.2.3
pytest-cov==2.11.1
requests==2.28.2
rsa==4.9
six==1.16.0
toml==0.10.2
urllib3==1.26.14
Werkzeug==2.2.2
yapf==0.32.0
zipp==3.13.0
